Rename unpack to retrieval, replace process_archives() with retrieval.process_elements()
